fix(android): use jint for domain register chain_id JNI param
jbyte is not in jni::sys imports; Kotlin chainId maps to jint.
Unblocks quic-jni build after domain_tx_signature FFI wiring.

+++ android/app/src/main/rust/quic-jni/src/identity_jni.rs
@@ -997,7 +997,7 @@ pub extern "system" fn Java_com_sovereignnetworkmobile_Identity_nativeBuildDomai
     content_mappings_json: JString<'local>,
     fee_payment_tx_hex: JString<'local>,
     metadata_json: JString<'local>,
-    chain_id: jbyte,
+    chain_id: jint,
 ) -> JString<'local> {
     if handle == 0 {
         return JString::default();
